Rework location assertion fees.
