+++ packages/onboarding/src/SolanaOnboarding.ts
@@ -1,24 +1,12 @@
 import { Cluster, Connection, PublicKey } from '@solana/web3.js'
-import * as AssertUtil from './AssertUtil'
+import * as HotspotOnboardingUtil from './HotspotOnboardingUtil'
 import OnboardingClient from './OnboardingClient'
-import { AnchorProvider, BN } from '@coral-xyz/anchor'
-import {
-  HNT_MINT,
-  sendAndConfirmWithRetry,
-  IOT_MINT,
-  MOBILE_MINT,
-  getAsset,
-} from '@helium/spl-utils'
+import { AnchorProvider } from '@coral-xyz/anchor'
+import { HNT_MINT, sendAndConfirmWithRetry, getAsset } from '@helium/spl-utils'
 import { init as initDc } from '@helium/data-credits-sdk'
-import {
-  init as initHem,
-  iotInfoKey,
-  keyToAssetKey,
-  mobileInfoKey,
-  rewardableEntityConfigKey,
-} from '@helium/helium-entity-manager-sdk'
-import { AssertData, DcProgram, HemProgram, HotspotType } from './types'
-import { daoKey, subDaoKey } from '@helium/helium-sub-daos-sdk'
+import { init as initHem, keyToAssetKey } from '@helium/helium-entity-manager-sdk'
+import { AssertData, DcProgram, DeviceType, HemProgram } from './types'
+import { daoKey } from '@helium/helium-sub-daos-sdk'
 import * as AssertMock from './__mocks__/AssertMock'
 import {
   createTransferInstructions,
@@ -89,15 +77,13 @@ export default class SolanaOnboarding {
     decimalGain,
     elevation,
     location,
-    hotspotTypes,
-    maker,
+    deviceType,
   }: {
     gateway: string
     decimalGain?: number
     elevation?: number
     location: string
-    maker?: PublicKey
-    hotspotTypes: HotspotType[]
+    deviceType: DeviceType
   }): Promise<AssertData> => {
     if (this.shouldMock) {
       return AssertMock.getAssertData()
@@ -106,7 +92,7 @@ export default class SolanaOnboarding {
     const dcProgram = await this.getDcProgram()
     const hemProgram = await this.getHemProgram()
 
-    return AssertUtil.getAssertData({
+    return HotspotOnboardingUtil.getAssertData({
       onboardingClient: this.onboardingClient,
       connection: this.connection,
       owner: this.wallet,
@@ -115,9 +101,8 @@ export default class SolanaOnboarding {
       decimalGain,
       gateway,
       elevation,
-      maker,
       nextLocation: location,
-      hotspotTypes,
+      deviceType,
       cluster: this.cluster,
     })
   }
@@ -173,26 +158,13 @@ export default class SolanaOnboarding {
     return results
   }
 
-  hotspotInfoToDetails = (value: {
-    asset: PublicKey
-    bumpSeed: number
-    isFullHotspot?: boolean
-    location: BN | null
-    numLocationAsserts?: number
-    elevation: number | null
-    gain: number | null
+  getHotspotDetails = async ({
+    networkType,
+    address,
+  }: {
+    address: string
+    networkType: 'MOBILE' | 'IOT'
   }) => {
-    const location = value.location?.toString('hex')
-    return {
-      elevation: value.elevation || undefined,
-      gain: value.gain || undefined,
-      location,
-      isFullHotspot: value.isFullHotspot,
-      numLocationAsserts: value.numLocationAsserts,
-    }
-  }
-
-  getHotspotDetails = async ({ type, address }: { address: string; type: 'MOBILE' | 'IOT' }) => {
     if (this.shouldMock) {
       return {
         elevation: 1,
@@ -203,37 +175,8 @@ export default class SolanaOnboarding {
       }
     }
 
-    try {
-      const hemProgram = await this.getHemProgram()
-
-      const mint = type === 'IOT' ? IOT_MINT : MOBILE_MINT
-      const subDao = subDaoKey(mint)[0]
-
-      const configKey = rewardableEntityConfigKey(subDao, type)
-
-      const entityConfig = await hemProgram.account.rewardableEntityConfigV0.fetchNullable(
-        configKey[0],
-      )
-      if (!entityConfig) return
-
-      if (type === 'IOT') {
-        const [info] = iotInfoKey(configKey[0], address)
-        const iotInfo = await hemProgram.account.iotHotspotInfoV0.fetch(info)
-        return this.hotspotInfoToDetails(iotInfo)
-      }
-
-      const [info] = mobileInfoKey(configKey[0], address)
-      const mobileInfo = await hemProgram.account.mobileHotspotInfoV0.fetch(info)
-      return this.hotspotInfoToDetails(mobileInfo)
-    } catch (e) {
-      const error = String(e)
-      if (error.startsWith('Error: Account does not exist or has no data')) {
-        // This is an expected error. It has not been onboarded to this network
-        return
-      }
-
-      throw e
-    }
+    const hemProgram = await this.getHemProgram()
+    return HotspotOnboardingUtil.getHotspotNetworkDetails({ networkType, address, hemProgram })
   }
 
   createTransferCompressedCollectableTxn = async ({
